moved repos to the same folder, fixed imports
