+++ Catacombs/client/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { BrowserRouter as Router } from "react-router-dom";
 import ApplicationViews from "./components/ApplicationViews";
-import { MovieProvider } from './components/Movies/MovieProvider';
-import { UserProvider } from './components/Users/UserProvider';
+import { MovieProvider } from './components/Repositories/MovieProvider';
+import { UserProvider } from './components/Repositories/UserProvider';
 import Header from './components/Header';
 
 function App() {
